+++ src/screens/UserDetail.tsx
@@ -4,7 +4,7 @@ import { View, Text } from 'react-native'
 const UserDetail = () => {
   return (
     <View>
-      <Text>UserList</Text>
+      <Text>UserDetail</Text>
     </View>
   )
 }